Add previewValue to DependencyKey, and fixed build errors on Xcode 13.
