+++ Examples/SpeechRecognition/SpeechRecognition/SpeechClient/Lorem.swift
@@ -1,0 +1,50 @@
+import ComposableArchitecture
+
+extension SpeechClient {
+  static var lorem: Self {
+    let isRecording = ActorIsolated(false)
+
+    return Self(
+      finishTask: { await isRecording.setValue(false) },
+      requestAuthorization: { .authorized },
+      startTask: { _ in
+          .init { c in
+            Task {
+              await isRecording.setValue(true)
+              var finalText = """
+              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor \
+              incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud \
+              exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute \
+              irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla \
+              pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui \
+              officia deserunt mollit anim id est laborum.
+              """
+              var text = ""
+              while await isRecording.value {
+                let word = finalText.prefix { $0 != " " }
+                try await Task.sleep(
+                  nanoseconds: UInt64(word.count) * NSEC_PER_MSEC * 50
+                  + .random(in: 0...(NSEC_PER_MSEC * 200))
+                )
+                finalText.removeFirst(word.count)
+                if finalText.first == " " {
+                  finalText.removeFirst()
+                }
+                text += word + " "
+                c.yield(
+                  .init(
+                    bestTranscription: .init(
+                      formattedString: text,
+                      segments: []
+                    ),
+                    isFinal: false,
+                    transcriptions: []
+                  )
+                )
+              }
+            }
+          }
+      }
+    )
+  }
+}
