+++ Examples/VoiceMemos/VoiceMemos/AudioPlayerClient/AudioPlayerClient.swift
@@ -1,18 +1,18 @@
 import Dependencies
 import Foundation
 
+struct AudioPlayerClient {
+  var play: @Sendable (URL) async throws -> Bool
+}
+
+enum AudioPlayerClientKey: DependencyKey {
+  static var previewValue = AudioPlayerClient.mock
+  static let testValue = AudioPlayerClient.unimplemented
+}
+
 extension DependencyValues {
   var audioPlayer: AudioPlayerClient {
     get { self[AudioPlayerClientKey.self] }
     set { self[AudioPlayerClientKey.self] = newValue }
   }
-
-  private enum AudioPlayerClientKey: LiveDependencyKey {
-    static let liveValue = AudioPlayerClient.live
-    static let testValue = AudioPlayerClient.unimplemented
-  }
-}
-
-struct AudioPlayerClient {
-  var play: @Sendable (URL) async throws -> Bool
 }
