+++ Examples/VoiceMemos/VoiceMemos/VoiceMemos.swift
@@ -233,44 +233,7 @@ struct VoiceMemos_Previews: PreviewProvider {
           ]
         ),
         reducer: VoiceMemos()
-          .dependency(\.audioPlayer, .mock)
-          .dependency(\.audioRecorder, .mock)
-          .dependency(\.openSettings) {}
       )
     )
-    .environment(\.colorScheme, .dark)
   }
-}
-
-extension AudioRecorderClient {
-  static var mock: Self {
-    let isRecording = ActorIsolated(false)
-    let currentTime = ActorIsolated(0.0)
-
-    return Self(
-      currentTime: { await currentTime.value },
-      requestRecordPermission: { true },
-      startRecording: { _ in
-        await isRecording.setValue(true)
-        while await isRecording.value {
-          try await Task.sleep(nanoseconds: NSEC_PER_SEC)
-          await currentTime.withValue { $0 += 1 }
-        }
-        return true
-      },
-      stopRecording: {
-        await isRecording.setValue(false)
-        await currentTime.setValue(0)
-      }
-    )
-  }
-}
-
-extension AudioPlayerClient {
-  static let mock = Self(
-    play: { _ in
-      try await Task.sleep(nanoseconds: NSEC_PER_SEC * 5)
-      return true
-    }
-  )
 }
