+++ Sources/Dependencies/Dependencies/Calendar.swift
@@ -1,18 +1,20 @@
 import Foundation
 import XCTestDynamicOverlay
 
-extension DependencyValues {
-  /// The current calendar that reducers should use when handling dates.
-  public var calendar: Calendar {
-    get { self[CalendarKey.self] }
-    set { self[CalendarKey.self] = newValue }
-  }
+#if swift(>=5.7)
+  extension DependencyValues {
+    /// The current calendar that reducers should use when handling dates.
+    public var calendar: Calendar {
+      get { self[CalendarKey.self] }
+      set { self[CalendarKey.self] = newValue }
+    }
 
-  private enum CalendarKey: LiveDependencyKey {
-    static let liveValue = Calendar.autoupdatingCurrent
-    static var testValue: Calendar {
-      XCTFail(#"Unimplemented: @Dependency(\.calendar)"#)
-      return .autoupdatingCurrent
+    private enum CalendarKey: LiveDependencyKey {
+      static let liveValue = Calendar.autoupdatingCurrent
+      static var testValue: Calendar {
+        XCTFail(#"Unimplemented: @Dependency(\.calendar)"#)
+        return .autoupdatingCurrent
+      }
     }
   }
-}
+#endif
